+++ src/tooling/piral-cli/src/common/browser.test.ts
@@ -2,7 +2,7 @@ import { openBrowser } from './browser';
 
 let error = false;
 
-jest.mock('opn', () => (...args) => {
+jest.mock('open', () => (...args) => {
   if (error) {
     throw new Error('Error occured');
   } else {
@@ -11,12 +11,14 @@ jest.mock('opn', () => (...args) => {
 });
 
 describe('Browser Module', () => {
-  it('opens browser successfully while not opening', async () => {
+  it('opens browser successfully', async () => {
     error = false;
     await openBrowser(false, 1234).then(result => expect(result).toBeUndefined());
     await openBrowser(true, 1234).then(result => expect(result).toBeUndefined());
+  });
+
+  it('handles errored opening', async () => {
     error = true;
     await openBrowser(true, 1234).then(result => expect(result).toBeUndefined());
-    error = false;
   });
 });